refactor text feature
