+++ crates/resvg/Cargo.toml
@@ -12,7 +12,7 @@ workspace = "../.."
 
 [[bin]]
 name = "resvg"
-required-features = ["svgz", "text", "text-layout", "system-fonts", "memmap-fonts"]
+required-features = ["svgz", "text", "system-fonts", "memmap-fonts"]
 
 [dependencies]
 bytemuck = "1.16"
@@ -31,14 +31,12 @@ once_cell = "1.21"
 png = "0.18.0"
 
 [features]
-default = ["svgz", "text", "text-layout", "system-fonts", "memmap-fonts", "raster-images"]
+default = ["svgz", "text", "system-fonts", "memmap-fonts", "raster-images"]
 # Enables SVGZ decoding
 svgz = ["usvg/svgz"]
-# Enables raw SVG text parsing support.
-text = ["usvg/text"]
 # Enables SVG text shaping and layout support.
 # Adds around 400KiB to your binary.
-text-layout = ["usvg/text-layout"]
+text = ["usvg/text"]
 # Enables system fonts loading (only for `text`).
 system-fonts = ["usvg/system-fonts"]
 # Enables font files memmaping for faster loading (only for `text`).
