+++ crates/usvg/Cargo.toml
@@ -15,7 +15,7 @@ workspace = "../.."
 
 [[bin]]
 name = "usvg"
-required-features = ["svgz", "text", "text-layout", "system-fonts", "memmap-fonts"]
+required-features = ["svgz", "text", "system-fonts", "memmap-fonts"]
 
 [dependencies]
 base64 = { version = "0.23", optional = true } # for embedded images
@@ -47,14 +47,12 @@ unicode-vo = { version = "0.1", optional = true }
 once_cell = "1.21"
 
 [features]
-default = ["svgz", "text", "text-layout", "system-fonts", "memmap-fonts"]
+default = ["svgz", "text", "system-fonts", "memmap-fonts"]
 # Enables SVGZ decoding
 svgz = ["flate2"]
-# Enables raw text parsing support.
-text = []
 # Enables text shaping and layout support.
 # Adds around 400KiB to your binary.
-text-layout = ["fontdb", "harfrust", "skrifa", "unicode-bidi", "unicode-script", "unicode-vo", "writer"]
+text = ["fontdb", "harfrust", "skrifa", "unicode-bidi", "unicode-script", "unicode-vo", "writer"]
 # Enables system fonts loading.
 system-fonts = ["fontdb/fs", "fontdb/fontconfig"]
 # Enables font files memmaping for faster loading.
