+++ crates/usvg/src/parser/text.rs
@@ -138,11 +138,11 @@ pub(crate) fn convert(
         stroke_bounding_box: dummy,
         abs_stroke_bounding_box: dummy,
         flattened: Box::new(Group::empty()),
-        #[cfg(feature = "text-layout")]
+        #[cfg(feature = "text")]
         layouted: vec![],
     };
 
-    #[cfg(feature = "text-layout")]
+    #[cfg(feature = "text")]
     if text::convert(&mut text, &state.opt.font_resolver, cache).is_none() {
         return;
     }
