+++ crates/usvg/src/tree/text.rs
@@ -30,6 +30,40 @@ impl Default for FontStretch {
     }
 }
 
+#[cfg(feature = "text")]
+impl From<fontdb::Stretch> for FontStretch {
+    fn from(stretch: fontdb::Stretch) -> Self {
+        match stretch {
+            fontdb::Stretch::UltraCondensed => FontStretch::UltraCondensed,
+            fontdb::Stretch::ExtraCondensed => FontStretch::ExtraCondensed,
+            fontdb::Stretch::Condensed => FontStretch::Condensed,
+            fontdb::Stretch::SemiCondensed => FontStretch::SemiCondensed,
+            fontdb::Stretch::Normal => FontStretch::Normal,
+            fontdb::Stretch::SemiExpanded => FontStretch::SemiExpanded,
+            fontdb::Stretch::Expanded => FontStretch::Expanded,
+            fontdb::Stretch::ExtraExpanded => FontStretch::ExtraExpanded,
+            fontdb::Stretch::UltraExpanded => FontStretch::UltraExpanded,
+        }
+    }
+}
+
+#[cfg(feature = "text")]
+impl From<FontStretch> for fontdb::Stretch {
+    fn from(stretch: FontStretch) -> Self {
+        match stretch {
+            FontStretch::UltraCondensed => fontdb::Stretch::UltraCondensed,
+            FontStretch::ExtraCondensed => fontdb::Stretch::ExtraCondensed,
+            FontStretch::Condensed => fontdb::Stretch::Condensed,
+            FontStretch::SemiCondensed => fontdb::Stretch::SemiCondensed,
+            FontStretch::Normal => fontdb::Stretch::Normal,
+            FontStretch::SemiExpanded => fontdb::Stretch::SemiExpanded,
+            FontStretch::Expanded => fontdb::Stretch::Expanded,
+            FontStretch::ExtraExpanded => fontdb::Stretch::ExtraExpanded,
+            FontStretch::UltraExpanded => fontdb::Stretch::UltraExpanded,
+        }
+    }
+}
+
 /// A font variation axis setting.
 ///
 /// Used for variable fonts to specify axis values like weight, width, etc.
@@ -78,6 +112,28 @@ impl Default for FontStyle {
     #[inline]
     fn default() -> FontStyle {
         Self::Normal
+    }
+}
+
+#[cfg(feature = "text")]
+impl From<fontdb::Style> for FontStyle {
+    fn from(style: fontdb::Style) -> Self {
+        match style {
+            fontdb::Style::Normal => FontStyle::Normal,
+            fontdb::Style::Italic => FontStyle::Italic,
+            fontdb::Style::Oblique => FontStyle::Oblique,
+        }
+    }
+}
+
+#[cfg(feature = "text")]
+impl From<FontStyle> for fontdb::Style {
+    fn from(style: FontStyle) -> Self {
+        match style {
+            FontStyle::Normal => fontdb::Style::Normal,
+            FontStyle::Italic => fontdb::Style::Italic,
+            FontStyle::Oblique => fontdb::Style::Oblique,
+        }
     }
 }
 
@@ -529,7 +585,7 @@ pub struct Text {
     pub(crate) stroke_bounding_box: Rect,
     pub(crate) abs_stroke_bounding_box: Rect,
     pub(crate) flattened: Box<Group>,
-    #[cfg(feature = "text-layout")]
+    #[cfg(feature = "text")]
     pub(crate) layouted: Vec<crate::layout::Span>,
 }
 
@@ -652,7 +708,7 @@ impl Text {
     /// This should only be used if you need more low-level access
     /// to the glyphs that make up the text. If you just need the
     /// outlines of the text, you should use `flattened` instead.
-    #[cfg(feature = "text-layout")]
+    #[cfg(feature = "text")]
     pub fn layouted(&self) -> &[crate::layout::Span] {
         &self.layouted
     }
